Convert Bolts tests to unit tests :raised_hands:
* Update integration tests to android-22
* Update tooling
