+++ Bolts/build.gradle
@@ -1,3 +1,5 @@
+import com.android.builder.core.BuilderConstants
+
 apply plugin: 'com.android.library'
 
 group = 'com.parse.bolts'
@@ -13,6 +15,10 @@ android {
         versionCode 1
         versionName project.version
     }
+}
+
+dependencies {
+    testCompile 'junit:junit:4.12'
 }
 
 android.libraryVariants.all { variant ->
@@ -35,7 +41,7 @@ android.libraryVariants.all { variant ->
         exclude '**/BuildConfig.class'
     }
 
-    if (name.equals(com.android.builder.core.BuilderConstants.RELEASE)) {
+    if (name.equals(BuilderConstants.RELEASE)) {
         artifacts.add('archives', task);
     }
 }
