+++ BoltsTest/build.gradle
@@ -17,5 +17,5 @@ android {
 dependencies {
     compile project(':Bolts')
 
-    androidTestCompile "com.android.support:support-v4:21.0.0"
+    androidTestCompile "com.android.support:support-v4:22.2.1"
 }
