+++ build.gradle
@@ -4,13 +4,19 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.0.0'
+        classpath 'com.android.tools.build:gradle:1.3.1'
+    }
+}
+
+allprojects {
+    repositories {
+        mavenCentral()
     }
 }
 
 ext {
     compileSdkVersion = 21
-    buildToolsVersion = "21.1.2"
+    buildToolsVersion = "22.0.1"
 
     minSdkVersion = 8
     targetSdkVersion = 21
